Cleaned up a mistake a made from the recent merge.
